frontend/explorer: only show combined search and terminal box, remove mini-terminal – #8337
